refactor(cli): hoist _console imports; shared action-row shape; snowdb.py -> root.py
Top-level _console imports everywhere (the fetch still happens at use time);
an autouse fixture restores the console pair after each CLI test so a --quiet
invocation can't leak global state; ingest/rasterize build their action rows
with the same single-comprehension shape doctor uses; the module holding the
top-level init/status commands is root.py now that the snowdb group is gone.
